agent-with-skills/app/requirements.txt
fastapi 
uvicorn 
llama-cpp-python 
python-dotenv
huggingface_hub
pydantic-settings
tenacity